# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gestion des Utilisateurs",
      "version": "1.0.0",
      "description": "API pour gérer les utilisateurs, y compris l'enregistrement, la connexion, et les opérations CRUD."
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
...
      "/users": {
        "get": {
          "summary": "Liste tous les utilisateurs",
          "operationId": "indexUsers",
          "parameters": [
            {
              "name": "role_id",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "status",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Liste des utilisateurs",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "nom": {
                          "type": "string"
                        },
                        "prenom": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Accès interdit"
            }
          }
        },
        "post": {
          "summary": "Créer un nouvel utilisateur",
          "operationId": "storeUser",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nom": {
                      "type": "string"
                    },
                    "prenom": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "password": {
                      "type": "string"
                    },
                    "role_id": {
                      "type": "integer"
                    },
                    "telephone": {
                      "type": "string"
                    },
                    "photo": {
                      "type": "string"
                    }
                  },
                  "required": ["nom", "prenom", "email", "password", "role_id", "telephone", "photo"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Utilisateur créé avec succès",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "nom": {
                        "type": "string"
                      },
                      "prenom": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string",
                        "format": "email"
                      }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Données invalides"
            }
          }
        }
      },
...
    }
  }